refactor topics screen into feature modules
